+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fixed: Do not spin forever if the spam filter eliminates too many transactions.
 - fixed: FIO OBT data not showing in received transaction memos
 - fixed: Return underLimit error for Simplex for very small quotes
+- removed: Removed pull-to-search feature from wallet list and transaction list scenes
 
 ## 4.11.0